scripts/init/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.7.14
cffi==1.17.1
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
minio==7.2.15
numpy==2.3.1
ollama==0.5.1
pip==24.3.1
portalocker==3.2.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
qdrant-client==1.15.0
sniffio==1.3.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
langchain-ollama==0.0.2
